+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Respect Long/Lat from Weathersettings for Sunrise/Sunset as well
+
 ## 1.1.40 (2022-10-14)
 
 * (theimo1221) Fix in Euro Heater to correctly set desired mode